+++ axelor-account/src/main/java/com/axelor/apps/account/service/FixedAssetService.java
@@ -17,8 +17,10 @@
  */
 package com.axelor.apps.account.service;
 
+import com.axelor.apps.account.db.AnalyticDistributionTemplate;
 import com.axelor.apps.account.db.FixedAsset;
 import com.axelor.apps.account.db.Invoice;
+import com.axelor.apps.account.db.MoveLine;
 import com.axelor.exception.AxelorException;
 import java.math.BigDecimal;
 import java.time.LocalDate;
@@ -52,5 +54,9 @@ public interface FixedAssetService {
    * @throws AxelorException
    */
   public void disposal(LocalDate disposalDate, BigDecimal disposalAmount, FixedAsset fixedAsset)
+      throws AxelorException;
+
+  public void createAnalyticOnMoveLine(
+      AnalyticDistributionTemplate analyticDistributionTemplate, MoveLine moveLine)
       throws AxelorException;
 }